main.go: add authorization header to read requests
Previously, the authorization header with the specified bearer token was
only set on remote-write requests and custom query requests. Now, with
the new authenticated Observatorium API, we want to test reading data
from an authenticated read API as well.
This commit ensures that the specified token is set on read requests.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>